# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API для управления арендой автомобилей (CRM).",
        "title": "Auto Hub API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/contracts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contracts"
                ],
                "summary": "Список договоров",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фильтр по статусу (ACTIVE, COMPLETED, CANCELLED)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ContractsListResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создаёт договор, меняет статус машины на RENTED и генерирует платёжный график",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contracts"
                ],
                "summary": "Создать договор",
                "parameters": [
                    {
                        "description": "Данные договора",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateContractRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Contract"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Машина не свободна",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/contracts/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает договор с машиной, водителем и платежами. Пересчитывает paid_amount.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contracts"
                ],
                "summary": "Детали договора",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID договора",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ContractDetailResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "При смене статуса с ACTIVE на другой — освобождает машину (FREE)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contracts"
                ],
                "summary": "Обновить договор",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID договора",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Поля для обновления",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateContractRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ContractResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contracts"
                ],
                "summary": "Удалить договор",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID договора",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/contracts/{id}/payments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contracts"
                ],
                "summary": "Платежи по договору",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID договора",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Payment"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.Car": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "inspection_before": {
                    "type": "string"
                },
                "mileage": {
                    "type": "integer"
                },
                "model": {
                    "type": "string"
                },
                "osago_before": {
                    "type": "string"
                },
                "plate_number": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.CarStatus"
                },
                "updated_at": {
                    "type": "string"
                },
                "vin": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "domain.CarStatus": {
            "type": "string",
            "enum": [
                "FREE",
                "RENTED",
                "REPAIR",
                "SOLD"
            ],
            "x-enum-varnames": [
                "CarStatusFree",
                "CarStatusRented",
                "CarStatusRepair",
                "CarStatusSold"
            ]
        },
        "domain.Contract": {
            "type": "object",
            "properties": {
                "car_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "driver_id": {
                    "type": "integer"
                },
                "end_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "monthly_payment": {
                    "type": "number"
                },
                "paid_amount": {
                    "type": "number"
                },
                "start_date": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.ContractStatus"
                },
                "total_amount": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.ContractStatus": {
            "type": "string",
            "enum": [
                "ACTIVE",
                "COMPLETED",
                "CANCELLED"
            ],
            "x-enum-varnames": [
                "ContractStatusActive",
                "ContractStatusCompleted",
                "ContractStatusCancelled"
            ]
        },
        "domain.Driver": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "license_num": {
                    "type": "string"
                },
                "passport_num": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.DriverStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.DriverStatus": {
            "type": "string",
            "enum": [
                "ACTIVE",
                "INACTIVE"
            ],
            "x-enum-varnames": [
                "DriverStatusActive",
                "DriverStatusInactive"
            ]
        },
...
        "domain.Payment": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "contract_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "due_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "paid_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.PaymentStatus"
                }
            }
        },
        "domain.PaymentStatus": {
            "type": "string",
            "enum": [
                "PAID",
                "UNPAID",
                "OVERDUE"
            ],
            "x-enum-varnames": [
                "PaymentStatusPaid",
                "PaymentStatusUnpaid",
                "PaymentStatusOverdue"
            ]
        },
...
        "dto.CarBrief": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "plate_number": {
                    "type": "string"
                }
            }
        },
...
        "dto.ContractDetailResponse": {
            "type": "object",
            "properties": {
                "car": {
                    "$ref": "#/definitions/domain.Car"
                },
                "car_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "driver": {
                    "$ref": "#/definitions/domain.Driver"
                },
                "driver_id": {
                    "type": "integer"
                },
                "end_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "monthly_payment": {
                    "type": "number"
                },
                "paid_amount": {
                    "type": "number"
                },
                "payments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Payment"
                    }
                },
                "remaining_balance": {
                    "type": "number"
                },
                "start_date": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.ContractStatus"
                },
                "total_amount": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "dto.ContractListItem": {
            "type": "object",
            "properties": {
                "car": {
                    "$ref": "#/definitions/dto.CarBrief"
                },
                "car_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "driver": {
                    "$ref": "#/definitions/dto.DriverName"
                },
                "driver_id": {
                    "type": "integer"
                },
                "end_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "monthly_payment": {
                    "type": "number"
                },
                "paid_amount": {
                    "type": "number"
                },
                "remaining_balance": {
                    "type": "number"
                },
                "start_date": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.ContractStatus"
                },
                "total_amount": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.ContractResponse": {
            "type": "object",
            "properties": {
                "car_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "driver_id": {
                    "type": "integer"
                },
                "end_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "monthly_payment": {
                    "type": "number"
                },
                "paid_amount": {
                    "type": "number"
                },
                "remaining_balance": {
                    "type": "number"
                },
                "start_date": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.ContractStatus"
                },
                "total_amount": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "dto.ContractsListResponse": {
            "type": "object",
            "properties": {
                "contracts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ContractListItem"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "dto.CreateContractRequest": {
            "type": "object",
            "required": [
                "car_id",
                "driver_id",
                "monthly_payment",
                "start_date",
                "total_amount"
            ],
            "properties": {
                "car_id": {
                    "type": "integer"
                },
                "driver_id": {
                    "type": "integer"
                },
                "monthly_payment": {
                    "type": "number"
                },
                "start_date": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "number"
                }
            }
        },
...
        "dto.DriverName": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateContractRequest": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string"
                },
                "monthly_payment": {
                    "type": "number"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "ACTIVE",
                        "COMPLETED",
                        "CANCELLED"
                    ]
                },
                "total_amount": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите токен в формате: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}